+++ src/main/java/com/mingles/metamingle/global/common/ApiStatus.java
@@ -1,4 +1,4 @@
-package com.mingles.metamingle.common;
+package com.mingles.metamingle.global.common;
 
 public enum ApiStatus {
     SUCCESS,
